+++ pkg/scaling/scale_handler.go
@@ -398,6 +398,8 @@ func buildScaler(ctx context.Context, client client.Client, triggerType string, 
 		return scalers.NewOpenstackSwiftScaler(ctx, config)
 	case "postgresql":
 		return scalers.NewPostgreSQLScaler(config)
+	case "predictkube":
+		return scalers.NewPredictKubeScaler(ctx, config)
 	case "prometheus":
 		return scalers.NewPrometheusScaler(config)
 	case "rabbitmq":