+++ CHANGELOG.md
@@ -25,6 +25,7 @@
 
 - Add New Relic Scaler ([#2387](https://github.com/kedacore/keda/pull/2387))
 - Add ActiveMQ Scaler ([#2305](https://github.com/kedacore/keda/pull/2305))
+- Add PredictKube Scaler ([#2418](https://github.com/kedacore/keda/pull/2418))
 
 ### Improvements